+++ examples/webhook/main.go
@@ -10,7 +10,7 @@ import (
 	"net/http"
 	"os"
 
-	api2convert "github.com/QaamGo/api2convert-go"
+	api2convert "github.com/QaamGo/api2convert-go/v10"
 )
 
 func main() {