+++ fileuploader_test.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 	"testing"
 
-	api2convert "github.com/QaamGo/api2convert-go"
-	"github.com/QaamGo/api2convert-go/internal/testutil"
+	api2convert "github.com/QaamGo/api2convert-go/v10"
+	"github.com/QaamGo/api2convert-go/v10/internal/testutil"
 )
 
 func stagedJob() api2convert.Job {