+++ wait_test.go
@@ -6,8 +6,8 @@ import (
 	"testing"
 	"time"
 
-	api2convert "github.com/QaamGo/api2convert-go"
-	"github.com/QaamGo/api2convert-go/internal/testutil"
+	api2convert "github.com/QaamGo/api2convert-go/v10"
+	"github.com/QaamGo/api2convert-go/v10/internal/testutil"
 )
 
 func TestWaitPollsUntilCompleted(t *testing.T) {