+++ webhook_test.go
@@ -5,8 +5,8 @@ import (
 	"strings"
 	"testing"
 
-	api2convert "github.com/QaamGo/api2convert-go"
-	"github.com/QaamGo/api2convert-go/internal/testutil"
+	api2convert "github.com/QaamGo/api2convert-go/v10"
+	"github.com/QaamGo/api2convert-go/v10/internal/testutil"
 )
 
 const whSecret = "whsec_test"